Correção da IDOrigem de IORIGEM para Long

+++ SIGFDomain/src/com/compsis/sigf/domain/Localizacao.java
@@ -115,9 +115,9 @@ public class Localizacao extends DomainObject implements Replicable, IOrigem{
 	}
 
 	@Override
-	public Integer getIdOrigem() {
+	public Long getIdOrigem() {
 		String idorigem = IDUtils.IdOrigem.getIdOrigem(this);
-		return idorigem!=null?Integer.parseInt(idorigem):null;
+		return idorigem!=null?Long.parseLong(idorigem):null;
 	}
 
 	@Override
@@ -146,9 +146,9 @@ public class Localizacao extends DomainObject implements Replicable, IOrigem{
 	}
 
 	@Override
-	public Integer getIdOrigemPai() {
+	public Long getIdOrigemPai() {
 		String idorigem = IDUtils.IdOrigem.getIdOrigem(this.getPraca());
-		return idorigem!=null?Integer.parseInt(idorigem):null;
+		return idorigem!=null?Long.parseLong(idorigem):null;
 	}
 
 	@Override
