+++ SIGFDomain/src/com/compsis/sigf/domain/Pista.java
@@ -140,15 +140,15 @@ public class Pista extends DomainObject implements Replicable, IOrigem{
 				}
 				
 				@Override
-				public Integer getIdOrigemPai() {
+				public Long getIdOrigemPai() {
 					String idorigem = IDUtils.IdOrigem.getIdOrigem(getThisInstanciaPista());
-					return idorigem!=null?Integer.parseInt(idorigem):null;
+					return idorigem!=null?Long.parseLong(idorigem):null;
 				}
 				
 				@Override
-				public Integer getIdOrigem() {
+				public Long getIdOrigem() {
 					String idorigem = IDUtils.IdOrigem.getIdOrigem(getThisInstanciaPista(), s);
-					return idorigem!=null?Integer.parseInt(idorigem):null;
+					return idorigem!=null?Long.parseLong(idorigem):null;
 				}
 				
 				@Override
@@ -311,9 +311,9 @@ public class Pista extends DomainObject implements Replicable, IOrigem{
     }
 
 	@Override
-	public Integer getIdOrigem() {
+	public Long getIdOrigem() {
 		String idorigem = IDUtils.IdOrigem.getIdOrigem(this);
-		return idorigem!=null?Integer.parseInt(idorigem):null;
+		return idorigem!=null?Long.parseLong(idorigem):null;
 	}
 
 	@Override
@@ -347,9 +347,9 @@ public class Pista extends DomainObject implements Replicable, IOrigem{
 	}
 
 	@Override
-	public Integer getIdOrigemPai() {
+	public Long getIdOrigemPai() {
 		String idorigem = IDUtils.IdOrigem.getIdOrigem(this.getLocalizacao());
-		return idorigem!=null?Integer.parseInt(idorigem):null;
+		return idorigem!=null?Long.parseLong(idorigem):null;
 	}
 
 	@Override
